matrix-alerts-automation/requirements.txt
boto3
slack_sdk
requests
certifi
aiohttp
python-dotenv